[🔨refactor] 메뉴 수정 기능 함수 분리 및 useInput 훅 사용
