+++ src/utils/MenuManagement/menuDB.ts
@@ -1,5 +1,5 @@
 import { getStorage, ref as storageRef, getDownloadURL, uploadBytes } from 'firebase/storage';
-import { addToDB, getCollection } from '../db';
+import { addToDB, getCollection, getDataFromDB, updateData } from '../db';
 import { MenuType } from '../../types/menuMangementType';
 
 const menuItemRef = getCollection('menuItem');
@@ -17,4 +17,20 @@ export const addMenu = async (menuInfo: MenuType, imageUrl: string) => {
 		...menuInfo,
 		imageUrl: imageUrl,
 	});
+};
+
+export const updateMenu = async (menuInfo: MenuType, newImageUrl: string) => {
+	const { id, name, price, category, imageName, imageUrl, soldout } = menuInfo;
+	const menuData = await getDataFromDB(menuItemRef, 'id', id);
+	if (menuData.docs.length !== 0) {
+		await updateData(menuData.docs[0].ref, {
+			id: id.toString(),
+			name,
+			price,
+			category,
+			imageName,
+			imageUrl: newImageUrl ? newImageUrl : imageUrl,
+			soldout,
+		});
+	}
 };
